% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{get_api}
\alias{get_api}
\title{Generic get and parse api request}
\usage{
get_api(url, api_user, api_token, simple = TRUE, query = NULL)
}
\arguments{
\item{api_token}{}
}
\value{

}
\description{
Generic get and parse api request
}
