% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{get_filenames}
\alias{get_filenames}
\title{Gets (probable) list of nc files for download}
\usage{
get_filenames(data_request, base_url = "")
}
\arguments{
\item{base_url}{}
}
\value{

}
\description{
Gets (probable) list of nc files for download
}
