% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{get_methods}
\alias{get_methods}
\title{Get methods for a particular sensor}
\usage{
get_methods(
  site,
  node,
  sensor,
  api_user,
  api_token,
  base_url = "https://ooinet.oceanobservatories.org/api/m2m/12576/sensor/inv"
)
}
\arguments{
\item{base_url}{}
}
\value{

}
\description{
Get methods for a particular sensor
}
