% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{get_sensor_info}
\alias{get_sensor_info}
\title{Get the information on a sensor}
\usage{
get_sensor_info(
  site,
  node,
  sensor,
  api_user,
  api_token,
  base_url = "https://ooinet.oceanobservatories.org/api/m2m/12586/vocab/inv"
)
}
\arguments{
\item{base_url}{}
}
\value{

}
\description{
Get the information on a sensor
}
