% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{get_sensor_location}
\alias{get_sensor_location}
\title{Get the location of the sensor}
\usage{
get_sensor_location(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Get the location of the sensor
}
