% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{get_sensor_meta}
\alias{get_sensor_meta}
\title{Get sensor metadata}
\usage{
get_sensor_meta(
  site,
  node,
  sensor,
  api_user,
  api_token,
  base_url = "https://ooinet.oceanobservatories.org/api/m2m/12576/sensor/inv",
  unpack_request = TRUE
)
}
\arguments{
\item{base_url}{}
}
\value{

}
\description{
Get sensor metadata
}
