% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{get_sensors}
\alias{get_sensors}
\title{Get names of sensors at a node}
\usage{
get_sensors(
  site,
  node,
  api_user,
  api_token,
  base_url = "https://ooinet.oceanobservatories.org/api/m2m/12576/sensor/inv"
)
}
\arguments{
\item{base_url}{}
}
\value{

}
\description{
Get names of sensors at a node
}
