% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{get_sites}
\alias{get_sites}
\title{Get list of sites}
\usage{
get_sites(
  api_user,
  api_token,
  base_url = "https://ooinet.oceanobservatories.org/api/m2m/12576/sensor/inv"
)
}
\arguments{
\item{base_url}{}
}
\value{

}
\description{
Get list of sites
}
