% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{get_streams}
\alias{get_streams}
\title{Get streams for a method}
\usage{
get_streams(
  site,
  node,
  sensor,
  method,
  api_user,
  api_token,
  base_url = "https://ooinet.oceanobservatories.org/api/m2m/12576/sensor/inv"
)
}
\arguments{
\item{base_url}{}
}
\value{

}
\description{
Get streams for a method
}
