% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{request_data}
\alias{request_data}
\title{Make a data request to the API Server}
\usage{
request_data(
  site,
  node,
  sensor,
  method,
  stream,
  api_user,
  api_token,
  query = NULL,
  base_url = "https://ooinet.oceanobservatories.org/api/m2m/12576/sensor/inv"
)
}
\arguments{
\item{base_url}{}
}
\value{

}
\description{
Make a data request to the API Server
}
