% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{unpack_metadata}
\alias{unpack_metadata}
\title{Unpack the contents of a api request for metadata}
\usage{
unpack_metadata(metadata)
}
\arguments{
\item{metadata}{metadata object returned by get_sensor_meta()}
}
\value{

}
\description{
Unpack the contents of a api request for metadata
}
